% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolker_ci.R
\name{bolker_ci}
\alias{bolker_ci}
\title{Confidence and prediction intervals from lme objects}
\usage{
bolker_ci(model, newdat, pred_int = FALSE, conf_level = 0.95)
}
\arguments{
\item{model}{The lme-model used for prediction.}

\item{newdat}{The data.frame to use for prediction.}

\item{pred_int}{Calculate prediction intervals (logical).}

\item{conf_level}{Confidence level for the calculation of both intervals.}
}
\value{
A datafame which has predictions, standarderrors, confidence and.
  prediction intervals
}
\description{
This function takes a mixed effect model (lme object) and a new dataframe as
  input and creates predictions, confidence intervals and prediction intervals
  using the method described by
  \href{https://bbolker.github.io/mixedmodels-misc/glmmFAQ.html#predictions-andor-confidence-or-prediction-intervals-on-predictions}{Ben Bolker}.
}
\examples{
# depends on an lme object from the nlme-package
m <- nlme::lme(mpg ~ disp + hp, random = ~1|cyl, data = mtcars)

# create new data.frame
new <- data.frame(disp = c(160, 170), hp = c(90, 100))
bolker_ci(m, new)
}
