% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_cis.R
\name{boot_cis}
\alias{boot_cis}
\title{Confidence intervals of bootMer objects}
\usage{
boot_cis(boot)
}
\arguments{
\item{boot}{bootMer object from which to retreive confidence interval}
}
\value{
A dataframe with confidence intervals
}
\description{
Confidence intervals of bootMer objects
}
\examples{
\dontrun{
boot_cis(bootMer.object)
}
}
