% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dta.R
\name{clean_dta}
\alias{clean_dta}
\title{Clean stata dataframes}
\usage{
clean_dta(df)
}
\arguments{
\item{df}{The dataframe, typically a file imported from stata.}
}
\value{
A datafame without attributes and all empty character cells ("") are
  converted to NA.
}
\description{
This function deletes all attributes from dataframes, especially those imported
from stata. Additionally, it converts empty character cells to NA.
}
\examples{
# needs sjlabelled and haven packages installed
clean_dta(mtcars)
}
