% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_ids.R
\name{distinct_ids}
\alias{distinct_ids}
\title{Count individual IDs}
\usage{
distinct_ids(df, id = id)
}
\arguments{
\item{df}{Dataframe.}

\item{id}{Column which is the id, default = id.}
}
\value{
Number of individual ids
}
\description{
Counts the number of unique ids within a dataframe
}
\examples{
mtcars$id <- 1:32
distinct_ids(mtcars)
#32
}
