% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhrisk_10y.R
\name{fhrisk_10y}
\alias{fhrisk_10y}
\title{Framingham Risk Calculator (10 Year Cardiovascular risk)}
\usage{
fhrisk_10y(
  gender,
  age,
  cholesterol,
  smoke,
  hdl,
  sysbp,
  bp_treated,
  mg_dl = TRUE
)
}
\arguments{
\item{gender}{"male" or "female"}

\item{age}{numeric}

\item{cholesterol}{numeric, default is mg/dL}

\item{smoke}{logical whether is smoking or not}

\item{hdl}{numeric, default is mg/dL}

\item{sysbp}{numeric, systolic blood pressure}

\item{bp_treated}{logical indicating whether received treatment for blood pressure}

\item{mg_dl}{default is TRUE, can be set to FALSE to calculate mmol/L}
}
\value{
character oject indicating 10% cardiovascular risk
}
\description{
This functions calculates the 10-year cardiovascular risk based on the Framingham Risk Score. I used data from 
\href{https://en.wikipedia.org/wiki/Framingham_Risk_Score}{Wikipedia}.
}
\examples{
fhrisk_10y(gender = "male", age = 60, cholesterol = 7, hdl = 1.2, sysbp = 120, bp_treated = TRUE, mg_dl = FALSE)

}
