% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_codes}
\alias{icd_codes}
\title{ICD 10 codes.}
\format{
An tibble with 16,058 rows and 7 variables
\describe{
  \item{icd_normcode}{ICD-10 code, detailed}
  \item{label}{Detailed ICD-10 label, corresponds to icd_normcode}
  \item{icd_code_short}{ICD-10 code, abbreviated to first three digits, useful if no detailed information is needed}
  \item{label_short}{Non-detailed ICD-10 label, corresponds to code}
  \item{icd_block_first}{first ICD-10 codes from current block of codes}
  \item{block_labe}{Label of current block}
  \item{chapter_label}{Name of ICD-10 chapter in which codes are organized}
}
}
\source{
\href{https://doi.org/10.5281/zenodo.2542833}{github}
}
\usage{
icd_codes
}
\description{
This dataset is derived from the the ICD10gm package and currently compiles a useful dataset from the 2018 version.
}
\references{
Donnachie (2020), Metadata Processing for the German Modification of the ICD-10 Coding System
(\href{https://edonnachie.github.io/ICD10gm}{github})
}
\keyword{datasets}
