% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable2.R
\name{kable2}
\alias{kable2}
\title{Knit table and round decimal places}
\usage{
kable2(df, digits = 2)
}
\arguments{
\item{df}{Dataframe you want to knit.}

\item{digits}{Number of decimal places, trailing zeros removed.}
}
\value{
Markdown table
}
\description{
Wrapper function to quickly knit a table and round numeric values.
}
\examples{
kable2(mtcars)

}
