% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kableone_nonnorm.R
\name{kableone_nonnorm}
\alias{kableone_nonnorm}
\title{Knit tableone with nonnormal values}
\usage{
kableone_nonnorm(x, nonnormal, ...)
}
\arguments{
\item{x}{Tableone object}

\item{nonnormal}{Vector containing nonnormal values for which median
is wanted.}

\item{...}{Further arguments passed to knitr::kable()}
}
\value{
Knitted version of a tableone object.
}
\description{
Knit tableone with nonnormal values
}
\examples{
library(tableone)
vars <- c("mpg", "cyl", "disp", "hp", "drat", "wt")
nonnormal <- c("mpg", "disp", "hp", "drat")
t1 <- tableone::CreateTableOne(vars, data = mtcars)
kableone_nonnorm(t1, nonnormal = nonnormal)

}
