% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_iqr.R
\name{median_iqr}
\alias{median_iqr}
\title{Calculate Median (Interquartile range) and return numeric.}
\usage{
median_iqr(df, variable)
}
\arguments{
\item{df}{Dataframe.}

\item{variable}{Numeric column used for calculation.}
}
\value{
Three new columns with median, q25, q75.
}
\description{
Calculates the median and interquartile range of a numeric column in a dataframe
and returns three numeric columns.
}
\examples{
median_iqr(mtcars, mpg)
#>   median    q25  q75
#> 1   19.2 15.425 22.8
}
