% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_iqr2.R
\name{median_iqr2}
\alias{median_iqr2}
\title{Calculate Median (Interquartile range) and return character.}
\usage{
median_iqr2(df, new_col, variable, digits = 1, to = " - ")
}
\arguments{
\item{df}{Dataframe.}

\item{new_col}{Name of the new column.}

\item{variable}{Numeric column used for calculation.}

\item{digits}{Optional number of decimal places.}

\item{to}{Optional character placed between Q25 and Q75, default is " - ".}
}
\value{
Summarised character column.
}
\description{
Calculates the median and interquartile range of a numeric column in a dataframe
and returns a character column in the form of median (Q25 - Q75).
}
\examples{
median_iqr2(mtcars, "mpg", mpg)
median_iqr2(mtcars, "mpg", mpg, to = " to ")
}
