% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mround.R
\name{mround}
\alias{mround}
\title{Round function for printing}
\usage{
mround(x, digits = 1)
}
\arguments{
\item{x}{Numeric object.}

\item{digits}{Number of digits after comma, default = 1.}
}
\value{
Rounded character object with trailing zeros.
}
\description{
This function can be used for tables, e.g. when trailing zeros are wanted.
}
\examples{

mround(10.123, digits = 1)
# "10.1"
mround(10.000, digits = 1)
# "10.0"

}
