% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_percent.R
\name{my_percent}
\alias{my_percent}
\title{Columnwise percentage calculation within pipe}
\usage{
my_percent(n, digits = 1)
}
\arguments{
\item{n}{Character object.}

\item{digits}{Number of decimal places to display.}
}
\value{
Column with percentage.
}
\description{
After counting or summarising in dplyr, generates percentages instead of
probabilites. It is a wrapper around round(n/sum(n)*100).
}
\examples{
mtcars \%>\%
  count(cyl) \%>\%
  mutate(p = my_percent(n))

}
