% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summary.R
\name{my_summary}
\alias{my_summary}
\title{Quick summaries command}
\usage{
my_summary(df, col, na.rm = TRUE)
}
\arguments{
\item{df}{Dataframe.}

\item{col}{Column to summarise.}

\item{na.rm}{Removes NAs before calculation, default is TRUE.}
}
\value{
Summarized dataframe.
}
\description{
Wrapper around the summarise command, returns
n = number of observations, min = minimum value, max = maximum value,
mean = arithmetric mean, median = median, p25 = 25-percentile,
p75 = 75-percentile.
}
