% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_pct.R
\name{n_pct}
\alias{n_pct}
\title{Prints number of observations and its percantage in brackets}
\usage{
n_pct(x, total, digits = 1, brackets = TRUE, trailing = TRUE)
}
\arguments{
\item{x}{Numerator}

\item{total}{Denominator}

\item{digits}{Digits to display after decimal point}

\item{brackets}{Whether percentages should be displayed in bracktes, alternative is seperation with ","}

\item{trailing}{Whether trailing zeroes should be displayed}
}
\value{

}
\description{
This function is especially useful when used in Rmarkdown documents or tables.
}
\examples{
n_pct(10, 1000)
# 10.0 (1.0\%)

n_pct(10, 1000, trailing = TRUE, brackets = FALSE)
# 10.0, 1.0\%

n_pct(10, 1000, trailing = FALSE, brackets = FALSE)
# 10, 1\%

}
