% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_iqr.R
\name{nice_iqr}
\alias{nice_iqr}
\title{Calculates and Interquartile ranges and prints them nicely}
\usage{
nice_iqr(x, to = TRUE, ...)
}
\arguments{
\item{x}{Vector to calculate IQR from}

\item{to}{Whether values should be separated by "to", alternative is "-"}

\item{...}{Arguments passed to bernr::comma()}
}
\value{
A character object
}
\description{
This function is especially useful when used in Rmarkdown documents
}
\examples{

set.seed(1)
x <- rnorm(10)
nice_iqr(x)
# "-0.5 to 0.6"

}
