% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_p.R
\name{nice_p}
\alias{nice_p}
\title{Create nice P-values}
\usage{
nice_p(pval, digits = 4)
}
\arguments{
\item{pval}{P-value to be converted}

\item{digits}{Number of digits to round}
}
\value{
A numeric object
}
\description{
This function returns nice p-values. This is useful especially for us with rmarkdown files and knitr::tables.
}
\examples{
set.seed(1)
x <- rnorm(100, mean = 0)
y <- rnorm(100, mean = 10)

t <- t.test(x, y)
nice_p(t$p.value)

}
