% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct.R
\name{pct}
\alias{pct}
\title{Calculates and nicely prints percent values based on two arguments}
\usage{
pct(x, total, digits = 1, trailing = TRUE)
}
\arguments{
\item{x}{Numerator}

\item{total}{Denominator}

\item{digits}{Digits to display after decimal point}

\item{trailing}{Whether trailing zeroes should be displayed}
}
\value{
A character object
}
\description{
This function is especially useful when used in Rmarkdown documents.
}
\examples{

pct(10, 100)
#"10.0\%"
pct(10, 100, trailing = FALSE)
#"10\%"
}
