% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_confint.R
\name{print_confint}
\alias{print_confint}
\title{Creates nice confidence report for tables}
\usage{
print_confint(df, new_col, estimate, lcl, ucl, digits = 1, to = " - ")
}
\arguments{
\item{df}{Dataframe.}

\item{new_col}{Name of the new column to create.}

\item{estimate}{Estimate for which confidence interval was created.}

\item{lcl}{Lower confidence value.}

\item{ucl}{Upper confidence value.}

\item{digits}{Round to digits.}

\item{to}{Separator used in the brackets.}
}
\description{
Creates a nice and convenient output for tables and graphs to report estimates
with their corresponding confidence intervals, e.g. 1.2 (95% CI 1.1 - 1.3).
}
\examples{
library(dplyr)
mtcars \%>\%
  summarise(mean = mean(mpg),
            n = n(),
            se = sd(mpg)/sqrt(n),
            lcl = mean - 1.96 * se,
            ucl = mean + 1.96 * se) \%>\%
            print_confint("mean_ci", estimate = mean, lcl = lcl,
                          ucl = ucl, digits = 2, to = " to ")
}
