% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_read.R
\name{pro_read}
\alias{pro_read}
\title{Read rds files from 'processed' folder}
\usage{
pro_read(file)
}
\arguments{
\item{file}{Name of the file.}
}
\description{
Reads rds files directly from the processed folder, which should be in
  the project's root folder.
}
\examples{
pro_read("dataframe.rds")
}
