% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_saver.R
\name{rds_saver}
\alias{rds_saver}
\title{Save rds files into folder}
\usage{
rds_saver(df, name, prefix = "00", postfix = "",
  folder = "processed")
}
\arguments{
\item{df}{Dataframe to save.}

\item{name}{Name of the file.}

\item{prefix}{Prefix before file, default = 00.}

\item{postfix}{Postfix after file, default = None.}

\item{folder}{Folder relative to project root.}
}
\description{
This is a convenience wrapper to write files into folders and append a prefix.
}
\examples{
rds_saver(mtcars, name = "mtcars", folder = "processed")

}
