% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shcs_to_rds.R
\name{shcs_to_rds}
\alias{shcs_to_rds}
\title{Parses .dta files from SHCS into .rds (batch)}
\usage{
shcs_to_rds(indir, outdir = "processed")
}
\arguments{
\item{indir}{Folder with SHCS data}

\item{outdir}{Folder to save rds data}
}
\value{
dta files in outdir folder
}
\description{
Parses .dta files from SHCS into .rds (batch)
}
\examples{

\dontrun{
shcs_to_rds(indir = "shcs_data", outdir = "processed")
}
}
