% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_last.R
\name{to_last}
\alias{to_last}
\title{Move column to the end of dataframe}
\usage{
to_last(df, ...)
}
\arguments{
\item{df}{Dataframe}

\item{...}{Columns that should be at the end of the Dataframe.}
}
\description{
Move column to the end of dataframe
}
\examples{
to_last(mtcars, cyl)
}
