% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_cis.R
\name{boot_cis}
\alias{boot_cis}
\title{Extract 95% CIs from bootMer objects}
\usage{
boot_cis(boot)
}
\arguments{
\item{boot}{Boot created with the bootMer command in the lme4 package.}
}
\value{
Dataframe with upper and lower confidence intervals.
}
\description{
Extract 95% CIs from bootMer objects
}
